criando auth pelo github

+++ server/src/routes/memories.ts
@@ -3,9 +3,16 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  /* Listagem de todas as Memórias */
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    request.jwtVerify()
+  })
+
+  // Listagem de todas as Memórias
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -20,7 +27,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  // Detalhe de uma Memória
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -33,9 +41,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
+  // Criação de uma Memória
   app.post('/memories', async (request) => {
     const bodySchema = z.object({
       content: z.string().uuid(),
@@ -50,14 +63,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '9109a215-d401-4c9e-be33-b90bb19d97d0',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  // Atualização de uma Memória
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -72,7 +86,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -86,12 +110,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  // Remover uma Memória
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
@@ -99,13 +134,4 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
   })
-
-  // // Detalhe de uma Memória
-  // app.get('/memories/:id', async () => {})
-  // // Criação de uma Memória
-  // app.post('memories', async () => {})
-  // // Atualização de uma Memória
-  // app.put('/memories/:id', async () => {})
-  // // Remover uma Memória
-  // app.delete('/memories/:id', async () => {})
 }
